modularised the leaderboard and calender from attendance
